[level 0] Title: 문자열 정렬하기 (1), Time: 1.11 ms, Memory: 69.3 MB -BaekjoonHub

+++ 프로그래머스/lv0/120850. 문자열 정렬하기 （1）/문자열 정렬하기 （1）.java
@@ -17,7 +17,7 @@ class Solution {
         answer = new int[list.size()];
         
         for (int i = 0 ; i < list.size() ; i++) {
-            answer[i] = list.get(i).intValue();
+            answer[i] = list.get(i);
         }
 
         // 배열 오름차순 정렬
